source/berkeley_humanoid_lite_lowlevel/requirements.txt
pyserial
python-can
cc.udp
onnxruntime
inputs
pynput
tables
